move dashboard handlers to routes package
